refactor: simplify app settings service
