+++ backend/src/main/java/org/booklore/service/appsettings/AppSettingService.java
@@ -1,11 +1,15 @@
 package org.booklore.service.appsettings;
 
+import lombok.extern.slf4j.Slf4j;
+import org.booklore.model.dto.request.MetadataRefreshOptions;
+import org.booklore.model.enums.MetadataProvider;
+import org.booklore.model.enums.MetadataReplaceMode;
+import org.booklore.repository.AppSettingsRepository;
 import org.springframework.transaction.annotation.Transactional;
 import org.booklore.config.AppProperties;
 import org.booklore.config.security.service.AuthenticationService;
 import org.booklore.exception.ApiError;
 import org.booklore.model.dto.BookLoreUser;
-import org.booklore.model.dto.request.MetadataRefreshOptions;
 import org.booklore.model.dto.settings.*;
 import org.booklore.model.entity.AppSettingEntity;
 import org.booklore.model.enums.AuditAction;
@@ -21,6 +25,7 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import tools.jackson.core.JacksonException;
 import tools.jackson.core.type.TypeReference;
+import tools.jackson.databind.ObjectMapper;
 
 import java.net.URI;
 import java.util.LinkedHashSet;
@@ -30,19 +35,22 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 @DependsOnDatabaseInitialization
 public class AppSettingService {
     private static final String DEFAULT_MOBILE_REDIRECT_URI = "grimmory://oauth2-callback";
     private static final String WILDCARD_REDIRECT_URI = "*";
 
     private final AppProperties appProperties;
-    private final SettingPersistenceHelper settingPersistenceHelper;
+    private final AppSettingsRepository appSettingsRepository;
+    private final ObjectMapper objectMapper;
     private final AuthenticationService authenticationService;
     private final AuditService auditService;
 
-    public AppSettingService(AppProperties appProperties, SettingPersistenceHelper settingPersistenceHelper, @Lazy AuthenticationService authenticationService, @Lazy AuditService auditService) {
+    public AppSettingService(AppProperties appProperties, AppSettingsRepository appSettingsRepository, ObjectMapper objectMapper, @Lazy AuthenticationService authenticationService, @Lazy AuditService auditService) {
         this.appProperties = appProperties;
-        this.settingPersistenceHelper = settingPersistenceHelper;
+        this.appSettingsRepository = appSettingsRepository;
+        this.objectMapper = objectMapper;
         this.authenticationService = authenticationService;
         this.auditService = auditService;
     }
@@ -70,13 +78,22 @@ public class AppSettingService {
             validateOidcForceOnlyMode(val);
         }
 
-        var setting = settingPersistenceHelper.appSettingsRepository.findByName(key.toString());
+        var setting = appSettingsRepository.findByName(key.toString());
+
         if (setting == null) {
             setting = new AppSettingEntity();
             setting.setName(key.toString());
         }
-        setting.setVal(settingPersistenceHelper.serializeSettingValue(key, val));
-        settingPersistenceHelper.appSettingsRepository.save(setting);
+
+        if (val == null) {
+            setting.setVal(null);
+        } else if (key.isJson()) {
+            setting.setVal(objectMapper.writeValueAsString(val));
+        } else {
+            setting.setVal(val.toString());
+        }
+
+        appSettingsRepository.save(setting);
 
         AuditAction action = switch (key) {
             case AppSettingKey k when k == AppSettingKey.OIDC_FORCE_ONLY_MODE -> AuditAction.OIDC_FORCE_ONLY_MODE_CHANGED;
@@ -179,10 +196,15 @@ public class AppSettingService {
         return buildPublicSetting();
     }
 
-    private Map<String, String> getSettingsMap() {
-        return settingPersistenceHelper.appSettingsRepository.findAll().stream()
+    private Map<AppSettingKey, String> getSettingsMap() {
+        return appSettingsRepository.findAll().stream()
                 .filter(entity -> entity.getName() != null && entity.getVal() != null)
-                .collect(Collectors.toMap(AppSettingEntity::getName, AppSettingEntity::getVal));
+                .collect(
+                    Collectors.toMap(
+                        (e) -> AppSettingKey.fromDbKey(e.getName()),
+                        AppSettingEntity::getVal
+                    )
+                );
     }
 
     private boolean isOIDCForceDisabled() {
@@ -193,18 +215,32 @@ public class AppSettingService {
         );
     }
 
+    private <T> T getJsonSetting(Map<AppSettingKey, String> settingsMap, AppSettingKey key, T defaultValue) {
+        String json = settingsMap.get(key);
+        if (json == null || json.isBlank()) {
+            return defaultValue;
+        }
+
+        try {
+            return objectMapper.readValue(json, new TypeReference<>() {});
+        } catch (JacksonException e) {
+            log.error("Failed to parse JSON for setting key '{}'. Using default value. Error: {}", key, e.getMessage());
+            return defaultValue;
+        }
+    }
+
     private PublicAppSetting buildPublicSetting() {
-        Map<String, String> settingsMap = getSettingsMap();
+        Map<AppSettingKey, String> settingsMap = getSettingsMap();
         PublicAppSetting.PublicAppSettingBuilder builder = PublicAppSetting.builder();
 
         builder.remoteAuthEnabled(appProperties.getRemoteAuth().isEnabled());
-        OidcProviderDetails details = settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.OIDC_PROVIDER_DETAILS, OidcProviderDetails.class, null, false);
+        OidcProviderDetails details = getJsonSetting(settingsMap, AppSettingKey.OIDC_PROVIDER_DETAILS, null);
         if (details != null) {
             details.setClientSecret(null);
         }
 
-        boolean oidcEnabled = Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.OIDC_ENABLED, "false"));
-        boolean oidcForceOnlyMode = Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.OIDC_FORCE_ONLY_MODE, "false"));
+        boolean oidcEnabled = Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.OIDC_ENABLED, "false"));
+        boolean oidcForceOnlyMode = Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.OIDC_FORCE_ONLY_MODE, "false"));
 
         if (isOIDCForceDisabled()) {
             oidcEnabled = false;
@@ -219,44 +255,35 @@ public class AppSettingService {
     }
 
     private AppSettings buildAppSettings() {
-        Map<String, String> settingsMap = getSettingsMap();
+        Map<AppSettingKey, String> settingsMap = getSettingsMap();
 
         AppSettings.AppSettingsBuilder builder = AppSettings.builder();
         builder.remoteAuthEnabled(appProperties.getRemoteAuth().isEnabled());
 
-        builder.defaultMetadataRefreshOptions(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.QUICK_BOOK_MATCH, MetadataRefreshOptions.class, settingPersistenceHelper.getDefaultMetadataRefreshOptions(), true));
-        builder.libraryMetadataRefreshOptions(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.LIBRARY_METADATA_REFRESH_OPTIONS, new TypeReference<>() {
-        }, List.of(), true));
-        builder.oidcProviderDetails(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.OIDC_PROVIDER_DETAILS, OidcProviderDetails.class, null, false));
-        builder.oidcRedirectUris(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.OIDC_REDIRECT_URIS, new TypeReference<>() {
-        }, List.of(DEFAULT_MOBILE_REDIRECT_URI), true));
-        builder.oidcAutoProvisionDetails(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.OIDC_AUTO_PROVISION_DETAILS, OidcAutoProvisionDetails.class, new OidcAutoProvisionDetails(), true));
-        builder.metadataProviderSettings(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.METADATA_PROVIDER_SETTINGS, MetadataProviderSettings.class, settingPersistenceHelper.getDefaultMetadataProviderSettings(), true));
-        builder.metadataMatchWeights(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.METADATA_MATCH_WEIGHTS, MetadataMatchWeights.class, settingPersistenceHelper.getDefaultMetadataMatchWeights(), true));
-        builder.metadataPersistenceSettings(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.METADATA_PERSISTENCE_SETTINGS, MetadataPersistenceSettings.class, settingPersistenceHelper.getDefaultMetadataPersistenceSettings(), true));
-        builder.metadataPublicReviewsSettings(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.METADATA_PUBLIC_REVIEWS_SETTINGS, MetadataPublicReviewsSettings.class, settingPersistenceHelper.getDefaultMetadataPublicReviewsSettings(), true));
-        builder.koboSettings(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.KOBO_SETTINGS, KoboSettings.class, settingPersistenceHelper.getDefaultKoboSettings(), true));
-        builder.coverCroppingSettings(settingPersistenceHelper.getJsonSetting(settingsMap, AppSettingKey.COVER_CROPPING_SETTINGS, CoverCroppingSettings.class, settingPersistenceHelper.getDefaultCoverCroppingSettings(), true));
-        builder.metadataProviderSpecificFields(
-            settingPersistenceHelper.getJsonSetting(
-                settingsMap,
-                AppSettingKey.METADATA_PROVIDER_SPECIFIC_FIELDS,
-                MetadataProviderSpecificFields.class,
-                settingPersistenceHelper.getDefaultMetadataProviderSpecificFields(),
-                true
-            )
-        );
-        builder.autoBookSearch(Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.AUTO_BOOK_SEARCH, "false")));
-        builder.uploadPattern(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.UPLOAD_FILE_PATTERN, "{authors}/<{series}/><{seriesIndex} - >{title}/{title}< - {authors}>< ({year})>"));
-        builder.similarBookRecommendation(Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.SIMILAR_BOOK_RECOMMENDATION, "true")));
-        builder.opdsServerEnabled(Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.OPDS_SERVER_ENABLED, "false")));
-        builder.komgaApiEnabled(Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.KOMGA_API_ENABLED, "false")));
-        builder.komgaGroupUnknown(Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.KOMGA_GROUP_UNKNOWN, "true")));
-        builder.pdfCacheSizeInMb(Integer.parseInt(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.PDF_CACHE_SIZE_IN_MB, "5120")));
-        builder.maxFileUploadSizeInMb(Integer.parseInt(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.MAX_FILE_UPLOAD_SIZE_IN_MB, "100")));
-        builder.metadataDownloadOnBookdrop(Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.METADATA_DOWNLOAD_ON_BOOKDROP, "true")));
+        builder.defaultMetadataRefreshOptions(getJsonSetting(settingsMap, AppSettingKey.QUICK_BOOK_MATCH, getDefaultMetadataRefreshOptions()));
+        builder.libraryMetadataRefreshOptions(getJsonSetting(settingsMap, AppSettingKey.LIBRARY_METADATA_REFRESH_OPTIONS, List.of()));
+        builder.oidcProviderDetails(getJsonSetting(settingsMap, AppSettingKey.OIDC_PROVIDER_DETAILS, null));
+        builder.oidcRedirectUris(getJsonSetting(settingsMap, AppSettingKey.OIDC_REDIRECT_URIS, List.of(DEFAULT_MOBILE_REDIRECT_URI)));
+        builder.oidcAutoProvisionDetails(getJsonSetting(settingsMap, AppSettingKey.OIDC_AUTO_PROVISION_DETAILS, new OidcAutoProvisionDetails()));
+        builder.metadataProviderSettings(getJsonSetting(settingsMap, AppSettingKey.METADATA_PROVIDER_SETTINGS, getDefaultMetadataProviderSettings()));
+        builder.metadataMatchWeights(getJsonSetting(settingsMap, AppSettingKey.METADATA_MATCH_WEIGHTS, getDefaultMetadataMatchWeights()));
+        builder.metadataPersistenceSettings(getJsonSetting(settingsMap, AppSettingKey.METADATA_PERSISTENCE_SETTINGS, getDefaultMetadataPersistenceSettings()));
+        builder.metadataPublicReviewsSettings(getJsonSetting(settingsMap, AppSettingKey.METADATA_PUBLIC_REVIEWS_SETTINGS, getDefaultMetadataPublicReviewsSettings()));
+        builder.koboSettings(getJsonSetting(settingsMap, AppSettingKey.KOBO_SETTINGS, getDefaultKoboSettings()));
+        builder.coverCroppingSettings(getJsonSetting(settingsMap, AppSettingKey.COVER_CROPPING_SETTINGS, getDefaultCoverCroppingSettings()));
+        builder.metadataProviderSpecificFields(getJsonSetting(settingsMap, AppSettingKey.METADATA_PROVIDER_SPECIFIC_FIELDS, getDefaultMetadataProviderSpecificFields()));
 
-        String sessionDurationStr = settingsMap.get(AppSettingKey.OIDC_SESSION_DURATION_HOURS.getDbKey());
+        builder.autoBookSearch(Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.AUTO_BOOK_SEARCH, "false")));
+        builder.uploadPattern(settingsMap.getOrDefault(AppSettingKey.UPLOAD_FILE_PATTERN, "{authors}/<{series}/><{seriesIndex} - >{title}/{title}< - {authors}>< ({year})>"));
+        builder.similarBookRecommendation(Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.SIMILAR_BOOK_RECOMMENDATION, "true")));
+        builder.opdsServerEnabled(Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.OPDS_SERVER_ENABLED, "false")));
+        builder.komgaApiEnabled(Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.KOMGA_API_ENABLED, "false")));
+        builder.komgaGroupUnknown(Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.KOMGA_GROUP_UNKNOWN, "true")));
+        builder.pdfCacheSizeInMb(Integer.parseInt(settingsMap.getOrDefault(AppSettingKey.PDF_CACHE_SIZE_IN_MB, "5120")));
+        builder.maxFileUploadSizeInMb(Integer.parseInt(settingsMap.getOrDefault(AppSettingKey.MAX_FILE_UPLOAD_SIZE_IN_MB, "100")));
+        builder.metadataDownloadOnBookdrop(Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.METADATA_DOWNLOAD_ON_BOOKDROP, "true")));
+
+        String sessionDurationStr = settingsMap.get(AppSettingKey.OIDC_SESSION_DURATION_HOURS);
         if (sessionDurationStr != null && !sessionDurationStr.isBlank()) {
             try {
                 builder.oidcSessionDurationHours(Integer.parseInt(sessionDurationStr));
@@ -264,8 +291,8 @@ public class AppSettingService {
             }
         }
 
-        boolean oidcEnabled = Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.OIDC_ENABLED, "false"));
-        boolean oidcForceOnlyMode = Boolean.parseBoolean(settingPersistenceHelper.getOrCreateSetting(AppSettingKey.OIDC_FORCE_ONLY_MODE, "false"));
+        boolean oidcEnabled = Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.OIDC_ENABLED, "false"));
+        boolean oidcForceOnlyMode = Boolean.parseBoolean(settingsMap.getOrDefault(AppSettingKey.OIDC_FORCE_ONLY_MODE, "false"));
 
         if (isOIDCForceDisabled()) {
             oidcEnabled = false;
@@ -275,8 +302,7 @@ public class AppSettingService {
         builder.oidcEnabled(oidcEnabled);
         builder.oidcForceOnlyMode(oidcForceOnlyMode);
 
-        builder.oidcGroupSyncMode(settingPersistenceHelper.getOrCreateSetting(
-                AppSettingKey.OIDC_GROUP_SYNC_MODE, "DISABLED"));
+        builder.oidcGroupSyncMode(settingsMap.getOrDefault(AppSettingKey.OIDC_GROUP_SYNC_MODE, "DISABLED"));
 
         builder.diskType(appProperties.getDiskType());
 
@@ -284,7 +310,7 @@ public class AppSettingService {
     }
 
     public String getSettingValue(String key) {
-        var setting = settingPersistenceHelper.appSettingsRepository.findByName(key);
+        var setting = appSettingsRepository.findByName(key);
         return setting != null ? setting.getVal() : null;
     }
 
@@ -294,12 +320,284 @@ public class AppSettingService {
     })
     @Transactional
     public void saveSetting(String key, String value) {
-        var setting = settingPersistenceHelper.appSettingsRepository.findByName(key);
+        var setting = appSettingsRepository.findByName(key);
         if (setting == null) {
             setting = new AppSettingEntity();
             setting.setName(key);
         }
         setting.setVal(value);
-        settingPersistenceHelper.appSettingsRepository.save(setting);
+        appSettingsRepository.save(setting);
+    }
+
+    private MetadataProviderSettings getDefaultMetadataProviderSettings() {
+        MetadataProviderSettings defaultMetadataProviderSettings = new MetadataProviderSettings();
+
+        MetadataProviderSettings.Amazon defaultAmazon = new MetadataProviderSettings.Amazon();
+        defaultAmazon.setEnabled(true);
+        defaultAmazon.setCookie(null);
+        defaultAmazon.setDomain("com");
+
+        MetadataProviderSettings.Google defaultGoogle = new MetadataProviderSettings.Google();
+        defaultGoogle.setEnabled(true);
+
+        MetadataProviderSettings.Goodreads defaultGoodreads = new MetadataProviderSettings.Goodreads();
+        defaultGoodreads.setEnabled(true);
+
+        MetadataProviderSettings.Hardcover defaultHardcover = new MetadataProviderSettings.Hardcover();
+        defaultHardcover.setEnabled(false);
+        defaultHardcover.setApiKey(null);
+
+        MetadataProviderSettings.Comicvine defaultComicvine = new MetadataProviderSettings.Comicvine();
+        defaultComicvine.setEnabled(false);
+        defaultComicvine.setApiKey(null);
+
+        MetadataProviderSettings.Douban defaultDouban = new MetadataProviderSettings.Douban();
+        defaultDouban.setEnabled(false);
+
+        MetadataProviderSettings.Ranobedb defaultRanobedb = new MetadataProviderSettings.Ranobedb();
+        defaultRanobedb.setEnabled(false);
+
+        defaultMetadataProviderSettings.setAmazon(defaultAmazon);
+        defaultMetadataProviderSettings.setGoogle(defaultGoogle);
+        defaultMetadataProviderSettings.setGoodReads(defaultGoodreads);
+        defaultMetadataProviderSettings.setHardcover(defaultHardcover);
+        defaultMetadataProviderSettings.setComicvine(defaultComicvine);
+        defaultMetadataProviderSettings.setRanobedb(defaultRanobedb);
+        defaultMetadataProviderSettings.setDouban(defaultDouban);
+
+        return defaultMetadataProviderSettings;
+    }
+
+    MetadataRefreshOptions getDefaultMetadataRefreshOptions() {
+        MetadataRefreshOptions.FieldProvider goodreadsGoogleProvider = MetadataRefreshOptions.FieldProvider.builder()
+                .p1(MetadataProvider.GoodReads)
+                .p2(MetadataProvider.Google)
+                .build();
+
+        MetadataRefreshOptions.FieldProvider nullProvider = MetadataRefreshOptions.FieldProvider.builder()
+                .build();
+
+        MetadataRefreshOptions.FieldOptions fieldOptions = MetadataRefreshOptions.FieldOptions.builder()
+                .title(goodreadsGoogleProvider)
+                .subtitle(goodreadsGoogleProvider)
+                .description(goodreadsGoogleProvider)
+                .authors(goodreadsGoogleProvider)
+                .publisher(goodreadsGoogleProvider)
+                .publishedDate(goodreadsGoogleProvider)
+                .seriesName(goodreadsGoogleProvider)
+                .seriesNumber(goodreadsGoogleProvider)
+                .seriesTotal(goodreadsGoogleProvider)
+                .isbn13(goodreadsGoogleProvider)
+                .isbn10(goodreadsGoogleProvider)
+                .language(goodreadsGoogleProvider)
+                .categories(goodreadsGoogleProvider)
+                .cover(goodreadsGoogleProvider)
+                .pageCount(goodreadsGoogleProvider)
+                .asin(nullProvider)
+                .goodreadsId(nullProvider)
+                .comicvineId(nullProvider)
+                .hardcoverId(nullProvider)
+                .hardcoverBookId(nullProvider)
+                .googleId(nullProvider)
+                .lubimyczytacId(nullProvider)
+                .amazonRating(nullProvider)
+                .amazonReviewCount(nullProvider)
+                .goodreadsRating(nullProvider)
+                .goodreadsReviewCount(nullProvider)
+                .hardcoverRating(nullProvider)
+                .hardcoverReviewCount(nullProvider)
+                .lubimyczytacRating(nullProvider)
+                .ranobedbId(nullProvider)
+                .ranobedbRating(nullProvider)
+                .audibleId(nullProvider)
+                .audibleRating(nullProvider)
+                .audibleReviewCount(nullProvider)
+                .moods(nullProvider)
+                .tags(nullProvider)
+                .build();
+
+        MetadataRefreshOptions.EnabledFields enabledFields = MetadataRefreshOptions.EnabledFields.builder()
+                .title(true)
+                .subtitle(true)
+                .description(true)
+                .authors(true)
+                .publisher(true)
+                .publishedDate(true)
+                .seriesName(true)
+                .seriesNumber(true)
+                .seriesTotal(true)
+                .isbn13(true)
+                .isbn10(true)
+                .language(true)
+                .categories(true)
+                .cover(true)
+                .pageCount(true)
+                .asin(true)
+                .goodreadsId(true)
+                .comicvineId(true)
+                .hardcoverId(true)
+                .hardcoverBookId(true)
+                .googleId(true)
+                .lubimyczytacId(true)
+                .amazonRating(true)
+                .amazonReviewCount(true)
+                .goodreadsRating(true)
+                .goodreadsReviewCount(true)
+                .hardcoverRating(true)
+                .hardcoverReviewCount(true)
+                .lubimyczytacRating(true)
+                .ranobedbId(false)
+                .ranobedbRating(false)
+                .audibleId(true)
+                .audibleRating(true)
+                .audibleReviewCount(true)
+                .moods(true)
+                .tags(true)
+                .build();
+
+        return MetadataRefreshOptions.builder()
+                .libraryId(null)
+                .refreshCovers(false)
+                .mergeCategories(true)
+                .reviewBeforeApply(false)
+                .replaceMode(MetadataReplaceMode.REPLACE_MISSING)
+                .fieldOptions(fieldOptions)
+                .enabledFields(enabledFields)
+                .build();
+    }
+
+    private MetadataMatchWeights getDefaultMetadataMatchWeights() {
+        return MetadataMatchWeights.builder()
+                .title(10)
+                .subtitle(1)
+                .description(10)
+                .authors(10)
+                .publisher(5)
+                .publishedDate(3)
+                .seriesName(2)
+                .seriesNumber(2)
+                .seriesTotal(1)
+                .isbn13(3)
+                .isbn10(5)
+                .language(2)
+                .pageCount(1)
+                .categories(10)
+                .amazonRating(3)
+                .amazonReviewCount(2)
+                .goodreadsRating(4)
+                .goodreadsReviewCount(2)
+                .hardcoverRating(2)
+                .hardcoverReviewCount(1)
+                .doubanRating(3)
+                .doubanReviewCount(2)
+                .ranobedbRating(2)
+                .lubimyczytacRating(2)
+                .audibleRating(0)
+                .audibleReviewCount(0)
+                .coverImage(5)
+                .build();
+    }
+
+    private MetadataPersistenceSettings getDefaultMetadataPersistenceSettings() {
+        MetadataPersistenceSettings.FormatSettings epubSettings = MetadataPersistenceSettings.FormatSettings.builder()
+                .enabled(false)
+                .maxFileSizeInMb(250)
+                .build();
+
+        MetadataPersistenceSettings.FormatSettings pdfSettings = MetadataPersistenceSettings.FormatSettings.builder()
+                .enabled(false)
+                .maxFileSizeInMb(250)
+                .build();
+
+        MetadataPersistenceSettings.FormatSettings cbxSettings = MetadataPersistenceSettings.FormatSettings.builder()
+                .enabled(false)
+                .maxFileSizeInMb(250)
+                .build();
+
+        MetadataPersistenceSettings.FormatSettings audiobookSettings = MetadataPersistenceSettings.FormatSettings.builder()
+                .enabled(false)
+                .maxFileSizeInMb(250)
+                .build();
+
+        MetadataPersistenceSettings.SaveToOriginalFile saveToOriginalFile = MetadataPersistenceSettings.SaveToOriginalFile.builder()
+                .epub(epubSettings)
+                .pdf(pdfSettings)
+                .cbx(cbxSettings)
+                .audiobook(audiobookSettings)
+                .build();
+
+        return MetadataPersistenceSettings.builder()
+                .saveToOriginalFile(saveToOriginalFile)
+                .convertCbrCb7ToCbz(false)
+                .moveFilesToLibraryPattern(false)
+                .build();
+    }
+
+    private MetadataPublicReviewsSettings getDefaultMetadataPublicReviewsSettings() {
+        return MetadataPublicReviewsSettings.builder()
+                .downloadEnabled(true)
+                .autoDownloadEnabled(false)
+                .providers(Set.of(
+                        MetadataPublicReviewsSettings.ReviewProviderConfig.builder()
+                                .provider(MetadataProvider.Amazon)
+                                .enabled(true)
+                                .maxReviews(5)
+                                .build(),
+                        MetadataPublicReviewsSettings.ReviewProviderConfig.builder()
+                                .provider(MetadataProvider.GoodReads)
+                                .enabled(false)
+                                .maxReviews(5)
+                                .build(),
+                        MetadataPublicReviewsSettings.ReviewProviderConfig.builder()
+                                .provider(MetadataProvider.Douban)
+                                .enabled(false)
+                                .maxReviews(5)
+                                .build()
+                ))
+                .build();
+    }
+
+    private KoboSettings getDefaultKoboSettings() {
+        return KoboSettings.builder()
+                .convertToKepub(true)
+                .conversionLimitInMb(100)
+                .convertCbxToEpub(false)
+                .conversionLimitInMbForCbx(100)
+                .conversionImageCompressionPercentage(85)
+                .forceEnableHyphenation(false)
+                .forwardToKoboStore(true)
+                .build();
+    }
+
+    private CoverCroppingSettings getDefaultCoverCroppingSettings() {
+        return CoverCroppingSettings.builder()
+                .verticalCroppingEnabled(false)
+                .horizontalCroppingEnabled(false)
+                .aspectRatioThreshold(2.5)
+                .smartCroppingEnabled(false)
+                .build();
+    }
+
+    private MetadataProviderSpecificFields getDefaultMetadataProviderSpecificFields() {
+        MetadataProviderSpecificFields fields = new MetadataProviderSpecificFields();
+        fields.setAsin(true);
+        fields.setAmazonRating(true);
+        fields.setAmazonReviewCount(true);
+        fields.setGoogleId(true);
+        fields.setGoodreadsId(true);
+        fields.setGoodreadsRating(true);
+        fields.setGoodreadsReviewCount(true);
+        fields.setHardcoverId(true);
+        fields.setHardcoverBookId(true);
+        fields.setHardcoverRating(true);
+        fields.setHardcoverReviewCount(true);
+        fields.setComicvineId(true);
+        fields.setLubimyczytacId(true);
+        fields.setLubimyczytacRating(true);
+        fields.setRanobedbRating(true);
+        fields.setAudibleId(true);
+        fields.setAudibleRating(true);
+        fields.setAudibleReviewCount(true);
+        return fields;
     }
 }
